Correct latitude value to be a Double

+++ OnTheMap/API/Response.swift
@@ -15,7 +15,7 @@ struct LocationResponse: Codable {
     let createdAt: String
     let firstName: String
     let lastName: String
-    let latitude: Float
+    let latitude: Double
     let longitude: Double
     let mapString: String
     let mediaURL: String
